requirements.txt
langchain~=0.1.0
openai~=1.6.1
chromadb~=0.4.22
langchainhub~=0.1.14
bs4~=0.0.1
tiktoken~=0.5.2
lxml~=5.1.0
beautifulsoup4~=4.12.2
soupsieve~=2.5
h11~=0.14.0
pip~=23.3.2
emoji~=2.9.0
attrs~=23.2.0
distro~=1.9.0
wheel~=0.41.2
filelock~=3.13.1
rsa~=4.9
pyasn1~=0.5.1
requests~=2.31.0
PyYAML~=6.0.1
idna~=3.6
multidict~=6.0.4
exceptiongroup~=1.2.0
sniffio~=1.3.0
tomli~=2.0.1
build~=1.0.3
click~=8.1.7
httpcore~=1.0.2
certifi~=2023.11.17
numpy~=1.26.3
setuptools~=68.2.0
overrides~=7.4.0
regex~=2023.12.25
sympy~=1.12
mpmath~=1.3.0
typer~=0.9.0
fsspec~=2023.12.2
aiohttp~=3.9.1
yarl~=1.9.4
cachetools~=5.3.2
urllib3~=2.1.0
packaging~=23.2
httpx~=0.26.0
pydantic~=2.5.3
anyio~=4.2.0
PyPika~=0.48.9
uvloop~=0.19.0
aiosignal~=1.3.1
frozenlist~=1.4.1
asgiref~=3.7.2
backoff~=2.2.1
fastapi~=0.108.0
starlette~=0.32.0.post1
websockets~=12.0
six~=1.16.0
python-dateutil~=2.8.2
posthog~=3.1.0
monotonic~=1.6
uvicorn~=0.25.0
httptools~=0.6.1
watchfiles~=0.21.0
onnxruntime~=1.16.3
kubernetes~=29.0.0
mmh3~=4.0.1
tenacity~=8.2.3
oauthlib~=3.2.2
Deprecated~=1.2.14
chardet~=5.2.0
langsmith~=0.0.78
SQLAlchemy~=2.0.25
jsonpatch~=1.33
websocket-client~=1.7.0
wrapt~=1.16.0
tokenizers~=0.15.0
humanfriendly~=10.0
coloredlogs~=15.0.1
marshmallow~=3.20.1
flatbuffers~=23.5.26
tqdm~=4.66.1
jsonpointer~=2.4
zipp~=3.17.0
pypdf~=3.17.4
python-dotenv~=1.0.0
pillow~=10.2.0
keras~=2.15.0
tensorflow~=2.15.0